All new Users now get added to the database, (will need to test once its in as ive only tested(thoroughly) on my cpanel.) New file get-recent-applications pulls that data along with a script that is currently in both index.html and admin-dashboard.html as a place holder till it works in the script.js.
fixed the application_date and follow_up_date so they are recorded along with the rest of the data.

Removed place holder information in index and admin and now replaced with a Id that fills out the recent apps.

Implented the reciept correctly so it formats equally to the others. (Some assisting code is being kept in to say if its connecting or not to help when testing once in the main app, will probably have another super small pr after to remove this.)